rename stepComponents -> step-components
